backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
ib_insync==0.9.86
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
nest-asyncio==1.5.8
numpy==1.26.3
pandas==2.1.4
python-multipart==0.0.6
